Move components to subfolder
