+++ components/home/Header.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { View } from 'react-native';
-import * as headerStyle from '../utils/stylesheets/header';
+import * as headerStyle from '../../utils/stylesheets/header';
 import { List, ListItem } from 'react-native-elements';
 
 export default class MainHeader extends Component {
