+++ components/home/ProfileHeader.js
@@ -2,8 +2,8 @@ import React, { Component } from 'react';
 import { View, Text } from 'react-native';
 import { Avatar } from 'react-native-elements';
 
-import * as profileHeaderStyle from '../utils/stylesheets/profileheader';
-import * as vars from '../utils/stylesheets/vars';
+import * as profileHeaderStyle from '../../utils/stylesheets/profileheader';
+import * as vars from '../../utils/stylesheets/vars';
 
 export default class ProfileHeader extends Component {
   render() {
